core-banking/src/com/example/banking/domain/CheckingAccount.java
package com.example.banking.domain;

/**
 * @author Binnur Kurt <binnur.kurt@gmail.com>
 */
public class CheckingAccount extends Account {
	private double overdraftAmount;

	public CheckingAccount(String iban, double balance, double overdraftAmount) {
		super(iban, balance);
		this.overdraftAmount = overdraftAmount;
	}

	public double getOverdraftAmount() {
		return overdraftAmount;
	}

	@Override
	public boolean withdraw(double amount) {
		System.out.println("CheckingAccount::withdraw");
		// validation
		if (amount <= 0.0)
			return false;
		// business rule
		if (amount > (balance + overdraftAmount))
			return false;
		this.balance -= amount;
		return true;
	}

}
